requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
serpapi==0.1.5
pymongo==4.6.2
python-dotenv==1.0.1
pydantic==2.6.4
